+++ src/main/java/net/brnbrd/delightful/common/item/DelightfulItems.java
@@ -102,6 +102,9 @@ public class DelightfulItems {
     public static final RegistryObject<Item> CHOPPED_CLOVER = registerItem("chopped_clover", () -> new ReverseTagItem((new Item.Properties()).food(Nutrition.CHOPPED_CLOVER), DelightfulItemTags.CLOVER));
     public static final RegistryObject<Item> CACTUS_FLESH = registerFood("cactus_flesh", Nutrition.CACTUS_FLESH);
     public static final RegistryObject<Item> CACTUS_STEAK = registerFood("cactus_steak", Nutrition.CACTUS_STEAK);
+    public static final RegistryObject<Item> CACTUS_CHILI = registerItem("cactus_chili", () -> new ConsumableItem((new Item.Properties()).food(Nutrition.CACTUS_CHILI).stacksTo(16).craftRemainder(Items.BOWL), true));
+    public static final RegistryObject<Item> CACTUS_SOUP = registerItem("cactus_soup", () -> new ConsumableItem((new Item.Properties()).food(Nutrition.CACTUS_SOUP).stacksTo(16).craftRemainder(Items.BOWL), true));
+    public static final RegistryObject<Item> CACTUS_SOUP_CUP = registerItem("cactus_soup_cup", () -> new CupItem((new Item.Properties()).food(Nutrition.CACTUS_SOUP_CUP), true, false));
     public static final RegistryObject<Item> FIELD_SALAD = registerItem("field_salad", () -> new ConsumableItem((new Item.Properties()).food(Nutrition.FIELD_SALAD).stacksTo(16).craftRemainder(Items.BOWL), true));
     public static final RegistryObject<Item> COCONUT_CURRY = registerItem("coconut_curry",
         () -> new CoconutCurryItem((new Item.Properties()).food(Nutrition.COCONUT_CURRY).stacksTo(16).craftRemainder(Items.BOWL)));