+++ changelog.md
@@ -4,11 +4,13 @@
 # 3.5.7
 
 ### Updates
+- Added Cactus Chili
+- Added Cactus Soup (with Cup for Miner's Delight)
 - Jade support for Mushroom Colonies, Salmonberries & Cantaloupes
 - Remove wood stripping recipes covered by Every Compat (highly recommend)
 - Added Cactus Chunk recipes from Habitat's cacti
-- Added "forge:cooked_nuts" tag, mainly for Vintage Delight compat
 - Hide disabled/conflicting items in JEI viewer
+- More compat recipes and tags
 
 ### Fixes
 - Fix Nature Staff & Zanite Knife (Aether) conflict